rename app to reg-scraper
