+++ .prettierrc.js
@@ -10,7 +10,7 @@ module.exports = {
     '<THIRD_PARTY_MODULES>',
     '^@api/(.*)$',
     '^@web/(.*)$',
-    '^@scraper/(.*)$',
+    '^@reg-scraper/(.*)$',
     '^@libs/(.*)$',
     '^[./]',
   ],
