+++ apps/reg-scraper/.env.template
