+++ apps/reg-scraper/.gitignore
