+++ apps/reg-scraper/Dockerfile
@@ -4,12 +4,12 @@ WORKDIR /usr/app
 ENV NODE_ENV production
 RUN npm install -g pnpm
 
-COPY dist/apps/scraper/.npmrc .
-COPY dist/apps/scraper/pnpm-lock.yaml .
-COPY dist/apps/scraper/package.json .
+COPY dist/apps/reg-scraper/.npmrc .
+COPY dist/apps/reg-scraper/pnpm-lock.yaml .
+COPY dist/apps/reg-scraper/package.json .
 RUN pnpm install -r --prod --prefer-offline
 
-COPY dist/apps/scraper .
+COPY dist/apps/reg-scraper .
 
 EXPOSE 3000
