+++ apps/reg-scraper/src/app.module.ts
@@ -4,11 +4,11 @@ import { ConfigModule, ConfigService } from '@nestjs/config'
 import { MongooseModule } from '@nestjs/mongoose'
 import { ScheduleModule } from '@nestjs/schedule'
 
-import { configuration } from '@scraper/config/configuration'
-import { OverrideModule } from '@scraper/override/override.module'
-import { QueueConsumerModule } from '@scraper/scraper/queue-consumer/queue-consumer.module'
-import { ScraperModule } from '@scraper/scraper/scraper.module'
-import { QueueStoreModule } from '@scraper/stores/queue-store/queue-store.module'
+import { configuration } from '@reg-scraper/config/configuration'
+import { OverrideModule } from '@reg-scraper/override/override.module'
+import { QueueConsumerModule } from '@reg-scraper/scraper/queue-consumer/queue-consumer.module'
+import { ScraperModule } from '@reg-scraper/scraper/scraper.module'
+import { QueueStoreModule } from '@reg-scraper/stores/queue-store/queue-store.module'
 
 import { AppController } from './app.controller'
 import { AppService } from './app.service'
