+++ apps/reg-scraper/src/app.service.ts
