+++ apps/reg-scraper/src/config/configuration.ts
