+++ apps/reg-scraper/src/main.ts
@@ -5,8 +5,8 @@ import { NestFactory } from '@nestjs/core'
 import { WinstonModule } from 'nest-winston'
 import * as winston from 'winston'
 
-import { Configuration, validateConfig } from '@scraper/config/configuration'
-import { ScraperService } from '@scraper/scraper/scraper.service'
+import { Configuration, validateConfig } from '@reg-scraper/config/configuration'
+import { ScraperService } from '@reg-scraper/scraper/scraper.service'
 
 import { AppModule } from './app.module'
 
@@ -17,7 +17,7 @@ async function bootstrap() {
       handleExceptions: true,
       handleRejections: true,
       defaultMeta: {
-        app: 'cgr-scraper',
+        app: 'reg-scraper',
       },
       transports: [
         new winston.transports.Console({
