+++ apps/reg-scraper/src/schema/course.schema.ts
