+++ apps/reg-scraper/src/schema/override.schema.ts
