+++ apps/reg-scraper/src/scraper/analytics/analyse.ts
