+++ apps/reg-scraper/src/scraper/params/course.param.ts
