+++ apps/reg-scraper/src/scraper/parser.ts
