+++ apps/reg-scraper/src/scraper/queue-consumer/queue-consumer.module.ts
@@ -2,9 +2,9 @@ import { BullModule } from '@nestjs/bull'
 import { Module } from '@nestjs/common'
 import { MongooseModule } from '@nestjs/mongoose'
 
-import { OverrideModule } from '@scraper/override/override.module'
-import { CourseSchema } from '@scraper/schema/course.schema'
-import { QueueStoreModule } from '@scraper/stores/queue-store/queue-store.module'
+import { OverrideModule } from '@reg-scraper/override/override.module'
+import { CourseSchema } from '@reg-scraper/schema/course.schema'
+import { QueueStoreModule } from '@reg-scraper/stores/queue-store/queue-store.module'
 
 import { QueueConsumerService } from './queue-consumer.service'
