+++ apps/reg-scraper/src/scraper/scraper.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common'
 import { MongooseModule } from '@nestjs/mongoose'
 
-import { OverrideModule } from '@scraper/override/override.module'
-import { CourseSchema } from '@scraper/schema/course.schema'
-import { QueueStoreModule } from '@scraper/stores/queue-store/queue-store.module'
+import { OverrideModule } from '@reg-scraper/override/override.module'
+import { CourseSchema } from '@reg-scraper/schema/course.schema'
+import { QueueStoreModule } from '@reg-scraper/stores/queue-store/queue-store.module'
 
 import { QueueProducerModule } from './queue-producer/queue-producer.module'
 import { ScraperController } from './scraper.controller'
