+++ apps/reg-scraper/src/scraper/selector/faculty.selector.ts
