+++ apps/reg-scraper/src/scraper/selector/search.selector.ts
