+++ apps/reg-scraper/src/scraper/types/CourseFetchJob.ts
