+++ apps/reg-scraper/src/stores/queue-store/queue-store.service.ts
