+++ apps/reg-scraper/tsconfig.app.json
