+++ apps/reg-scraper/tsconfig.json
