+++ apps/reg-scraper/tsconfig.spec.json
