+++ workspace.json
@@ -6,7 +6,7 @@
     "course-utils": "libs/course-utils",
     "api": "apps/api",
     "web": "apps/web",
-    "scraper": "apps/scraper",
+    "reg-scraper": "apps/reg-scraper",
     "web-e2e": "apps/web-e2e"
   }
 }
